+++ apps/playground/CHANGELOG.md
@@ -1,5 +1,19 @@
 # playground
 
+## 1.0.0-beta.126
+
+### Patch Changes
+
+- Updated dependencies [22c2a64]
+  - sit-onyx@1.0.0-beta.126
+
+## 1.0.0-beta.125
+
+### Patch Changes
+
+- Updated dependencies [1d803fb]
+  - sit-onyx@1.0.0-beta.125
+
 ## 1.0.0-beta.124
 
 ### Patch Changes