+++ packages/sit-onyx/CHANGELOG.md
@@ -1,5 +1,17 @@
 # sit-onyx
 
+## 1.0.0-beta.126
+
+### Patch Changes
+
+- 22c2a64: fix(OnyxFormElement): fix broken counter text when maxlength with counter is used
+
+## 1.0.0-beta.125
+
+### Patch Changes
+
+- 1d803fb: OnyxCard: increase border radius
+
 ## 1.0.0-beta.124
 
 ### Minor Changes